_old/pyproject.toml
[tool.poetry]
name = "awesome"
version = "0.1.0"
description = ""
authors = ["Jean-Michel Girard <Jean.M.Girard@Outlook.com>"]
readme = "README.md"
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.12"
mkdocs-git-latest-changes-plugin = "^0.0.20"
mkdocs-include-dir-to-nav = "^1.2.0"
mkdocs-bootstrap = "^1.1.1"
mkdocs = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
wikilinks = "^0.0.4"
smarty = "^0.3.3"
pymdown-extensions = "^10.21"
pre-commit = "^3.7.0"
libsass = "^0.23.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.1.1"

[tool.poetry.group.docs]
optional = true

# Python-style entrypoints and scripts are easily expressed
[tool.poetry.scripts]
serve = "mkdocs serve"
